+++ R/modelling.R
@@ -55,7 +55,10 @@
 #'                options = list(max_iterations = 3, feature.minfreq = 5, 
 #'                               c1 = 0, c2 = 1)) 
 #'   model
-#'   stats <- summary(model, "modeldetails.txt")
+#'   weights <- coefficients(model)
+#'   head(weights$states, n = 20)
+#'   head(weights$transitions, n = 20)
+#'   stats   <- summary(model, "modeldetails.txt")
 #'   stats
 #'   plot(stats$iterations$loss)
 #' 
@@ -103,6 +106,7 @@
 crf <- function(x, y, group, 
                 method = c("lbfgs", "l2sgd", "averaged-perceptron", "passive-aggressive", "arow"), 
                 options = crf_options(method)$default, 
+                embeddings,
                 file = "annotator.crfsuite", trace = FALSE, FUN = identity, ...){
   type <- "crf1d"
   file <- normalizePath(path.expand(file), mustWork = FALSE)
@@ -123,10 +127,19 @@ crf <- function(x, y, group,
     cat(sprintf("CRFsuite training progress logged to file %s", f), sep = "\n")
   }
   on.exit(file.remove(f))
+  if(missing(embeddings)){
+    embeddings <- matrix(numeric(), nrow = 0, ncol = 0)
+  }else{
+    embeddings <- as.matrix(embeddings)
+    stopifnot(nrow(embeddings) == nrow(x))
+    if(is.null(colnames(embeddings))){
+      colnames(embeddings) <- sprintf("embedding_dim%s", seq_len(ncol(embeddings)))
+    }
+  }
   model <- crfsuite_model_build(file_model = file, 
-                       doc_id = group, y = y, x = x,
-                       options = options, method = method, type = type,
-                       trace = trace)
+                                doc_id = group, y = y, x = x, embeddings = embeddings,
+                                options = options, method = method, type = type,
+                                trace = trace)
   model$attribute_names <- colnames(x)
   model$log <- readLines(f)
   model$FUN <- FUN
@@ -174,7 +187,7 @@ print.crf <- function(x, ...){
 #' @export
 summary.crf <- function(object, file, ...){
   stopifnot(file.exists(object$file_model))
-
+  
   out <- list()
   out$active <- list()
   out$active$features <- as.numeric(gsub("(^Number of active features: )(.+) (.+)$", "\\2", grep("^Number of active features: ", object$log, value = TRUE)))
@@ -262,7 +275,7 @@ summary.crf <- function(object, file, ...){
 #' file.remove(udmodel$file)
 #' \dontshow{\} # End of main if statement running only if the required packages are installed}
 #' }
-predict.crf <- function(object, newdata, group, type = c("marginal", "sequence"), trace = FALSE, ...){
+predict.crf <- function(object, newdata, embeddings, group, type = c("marginal", "sequence"), trace = FALSE, ...){
   stopifnot(file.exists(object$file_model))
   trace <- as.integer(trace)
   newdata <- as.matrix(newdata)
@@ -276,7 +289,16 @@ predict.crf <- function(object, newdata, group, type = c("marginal", "sequence")
   }else{
     recodegroup <- FALSE
   }
-  scores <- crfsuite_predict(file_model = object$file_model, doc_id = group, x = newdata, trace = trace)
+  if(missing(embeddings)){
+    embeddings <- matrix(numeric(), nrow = 0, ncol = 0)
+  }else{
+    embeddings <- as.matrix(embeddings)
+    stopifnot(nrow(embeddings) == nrow(newdata))
+    if(is.null(colnames(embeddings))){
+      colnames(embeddings) <- sprintf("embedding_dim%s", seq_len(ncol(embeddings)))
+    }
+  }
+  scores <- crfsuite_predict(file_model = object$file_model, doc_id = group, x = newdata, embeddings = embeddings, trace = trace)
   if(type == "marginal"){
     scores <- scores$viterbi
   }else if(type == "sequence"){