+++ src/rcpp_crfsuite.cpp
@@ -43,6 +43,7 @@ Rcpp::List crfsuite_model_build(const char* file_model,
                                 const std::vector<int> doc_id, 
                                 const std::vector<std::string> y, 
                                 Rcpp::CharacterMatrix x,
+                                Rcpp::NumericMatrix embeddings,
                                 Rcpp::List options,
                                 const std::string method = "lbfgs", 
                                 const std::string type = "crf1d",
@@ -85,6 +86,15 @@ Rcpp::List crfsuite_model_build(const char* file_model,
       for(int j=0; j<nfeatures; j++) {  
         if(!Rcpp::CharacterVector::is_na(x(i, j))){
           termfeatures.push_back(CRFSuite::Attribute(std::string(x(i, j)), 1.));  
+        }
+      }
+      if(embeddings.nrow() > 0){
+        Rcpp::CharacterVector fields = colnames(embeddings);
+        for(int j=0; j<(embeddings.ncol()); j++) {  
+          std::string field = Rcpp::as<std::string>(fields[j]);
+          if(!Rcpp::NumericVector::is_na(embeddings(i, j))){
+            termfeatures.push_back(CRFSuite::Attribute(field, embeddings(i, j)));  
+          }
         }
       }
       termsequenceattributes.push_back(termfeatures);
@@ -166,6 +176,7 @@ Rcpp::List crfsuite_model_coefficients(const char* file_model){
 Rcpp::List crfsuite_predict(const std::string file_model, 
                             const std::vector<int> doc_id, 
                             Rcpp::CharacterMatrix x,
+                            Rcpp::NumericMatrix embeddings,
                             int trace = 0){
   std::vector<std::string> labels;
   std::vector<double> marginal;
@@ -201,6 +212,15 @@ Rcpp::List crfsuite_predict(const std::string file_model,
       for(int j=0; j<nfeatures; j++) {  
         if(!Rcpp::CharacterVector::is_na(x(i, j))){
           termfeatures.push_back(CRFSuite::Attribute(std::string(x(i, j)), 1.));
+        }
+      }
+      if(embeddings.nrow() > 0){
+        Rcpp::CharacterVector fields = colnames(embeddings);
+        for(int j=0; j<(embeddings.ncol()); j++) {  
+          std::string field = Rcpp::as<std::string>(fields[j]);
+          if(!Rcpp::NumericVector::is_na(embeddings(i, j))){
+            termfeatures.push_back(CRFSuite::Attribute(field, embeddings(i, j)));  
+          }
         }
       }
       termsequenceattributes.push_back(termfeatures);